"Allow forget password route to accept GET requests and send password reset email; update user email in database on update\_user request"
This commit updates the `forget_password` route to accept both POST and GET requests. When a GET request is received, it searches for a user with the provided email address in both the `usher` and `planner` tables. If a user is found, a password reset token is generated and an email is sent to the user with a link to reset their password.

Additionally, the `update_user` function has been updated to set the user's email to `bernadettestella27@gmail.com` when updating their information in the database.